clean up a bunch of things
