+++ package.json
@@ -27,7 +27,6 @@
     "owasp-password-strength-test": "^1.3.0",
     "react": "^16.8.1",
     "react-dom": "^16.8.1",
-    "react-loadable": "^5.5.0",
     "react-reactive-form": "^1.0.30",
     "react-redux": "^7.0.3",
     "react-router-dom": "^5.0.0",
@@ -42,7 +41,6 @@
     "@types/owasp-password-strength-test": "^1.3.0",
     "@types/react": "^16.8.17",
     "@types/react-dom": "^16.8.4",
-    "@types/react-loadable": "^5.5.1",
     "@types/react-redux": "^7.0.8",
     "@types/react-router-dom": "^4.3.3",
     "@types/reactstrap": "^8.0.1",
@@ -55,6 +53,7 @@
     "source-map-loader": "^0.2.4",
     "style-loader": "^0.23.1",
     "tslint": "^5.16.0",
+    "tslint-eslint-rules": "^5.4.0",
     "typescript": "^3.4.5",
     "uglifyjs-webpack-plugin": "^2.1.2",
     "webpack-cli": "^3.3.2",
