+++ src/adventure-creator/selected-state/scss/SelectedState.scss
@@ -43,7 +43,7 @@ $maxWidth: 400px;
     @include grayBackgroundWhenDisabled;
     background: $green;
     border: 1px solid $gray;
-    color: #fcfcfc;
+    color: $offWhite;
     font-size: 16px;
     margin-bottom: 10px;
     padding: 10px 20px;
@@ -51,7 +51,7 @@ $maxWidth: 400px;
 
     &:hover {
       background: $blue;
-      color: #fcfcfc;
+      color: $offWhite;
       cursor: pointer;
     }
