+++ src/app/scss/Header.scss
@@ -2,7 +2,7 @@
 @import 'src/scss/variables.scss';
 
 .header {
-  background: $themeBlue; // $blue
+  background: $themeBlue;
   box-shadow: 0 0 8px 4px rgba(0,0,0,0.3);
   font-family: 'Open Sans', sans-serif;
   font-weight: bold;
@@ -15,19 +15,22 @@
   width: 100%;
   z-index: 1;
 
-  &__title {
-    @include resetButton;
+  &__logo {
     @include undecoratedLink;
+    align-items: center;
     color: rgb(255,255,255);
     cursor: pointer;
-    align-items: center;
-    font-size: 40px; // 32px
-    font-weight: 900;
-    margin: 3px; // 8px
-    padding: 0px 20px; // 2px 4px
+    display: flex;
+    justify-content: center;
+    padding: 10px;
 
     &:hover {
-      color: #fcfcfc;
+      color: $offWhite;
+    }
+
+    &__img {
+      height: 40px;
+      width: auto;
     }
   }
 
@@ -42,14 +45,14 @@
     @include undecoratedLink;
     border: 2px solid white;
     border-radius: 5px;
-    color: #fcfcfc;
-    font-size: 14px; //18px
-    margin: 0 6px; // 0 6px
-    padding: 8px 20px; // 10px 20px
+    color: $offWhite;
+    font-size: 16px;
+    margin: 0 6px;
+    padding: 8px 20px;
     white-space: nowrap;
 
     &:hover {
-      background: #fcfcfc;
+      background: $offWhite;
       color: $themeBlue;
       cursor: pointer;
     }
