+++ src/game/scss/Game.scss
@@ -16,10 +16,7 @@
   }
 
   &__content {
-    background-color: black;
-    background-position: center;
-    background-repeat: no-repeat;
-    background-size: contain;
+    background: black center / contain no-repeat;
     bottom: 0;
     left: 0;
     position: absolute;
