+++ src/landing-page/LandingPage.tsx
@@ -4,6 +4,16 @@ const styles = require('./scss/LandingPage.scss') // tslint:disable-line no-var-
 
 const LandingPage = () => (
   <div className={styles.landingPage}>
+    <div className={styles.landingPage__info}>
+      <h1 className={styles.landingPage__info__header}>
+        Are your kids being safe online?
+      </h1>
+      <p className={styles.landingPage__info__text}>
+        {/* tslint:disable-next-line:max-line-length */}
+        We will educate, engage, and prevent dangerous online situations for your children through our interactive programs on cybersecurity.
+      </p>
+      <button className={styles.landingPage__info__button} />
+    </div>
   </div>
 )
