--- src/loading-page/scss/LoadingPage.scss
@@ -1,9 +1,0 @@
-@import 'src/scss/variables.scss';
-
-.loadingPage {
-  align-items: center;
-  color: $blue;
-  display: flex;
-  justify-content: center;
-  padding: 2rem;
-}
