+++ src/login/scss/Login.scss
@@ -1,4 +1,5 @@
 @import 'src/scss/mixins.scss';
+@import 'src/scss/variables.scss';
 
 // TODO create mixin(s) for shared scss (between login/sign up, and possibly others - buttons?)
 
@@ -14,7 +15,7 @@
       @include grayBackgroundWhenDisabled;
       background: $green;
       border-radius: 5px;
-      color: #fcfcfc;
+      color: $offWhite;
       font-size: 18px;
       margin: 0 2px;
       padding: 10px 20px;
@@ -22,7 +23,7 @@
 
       &:hover {
         background: $blue;
-        color: #fcfcfc;
+        color: $offWhite;
         cursor: pointer;
       }
     }
