+++ src/minigames/PasswordList/scss/PasswordList.scss
@@ -74,7 +74,7 @@
     @include resetButton;
     background: $blue;
     border-radius: 5px;
-    color: #fcfcfc;
+    color: $offWhite;
     font-size: 16px;
     margin-top: 10px;
     max-width: 200px;
@@ -83,7 +83,7 @@
 
     &:hover {
       background: $green;
-      color: #fcfcfc;
+      color: $offWhite;
       cursor: pointer;
     }
   }
@@ -94,7 +94,7 @@
   @include resetButton;
   background: $blue;
   border-radius: 5px;
-  color: #fcfcfc;
+  color: $offWhite;
   font-size: 16px;
   margin-top: 10px;
   max-width: 200px;
@@ -103,7 +103,7 @@
 
   &:hover {
     background: $green;
-    color: #fcfcfc;
+    color: $offWhite;
     cursor: pointer;
   }
 }
