+++ src/notification-buttons/notification-button/scss/NotificationButton.scss
@@ -1,4 +1,5 @@
 .notificationButton {
+  // background: transparent center / contain no-repeat;
   background-color: transparent !important;
   background-position: center;
   background-repeat: no-repeat;
