+++ src/scss/variables.scss
@@ -1,4 +1,5 @@
 // colors
+$offWhite: rgb(252,252,252);
 $lightGray: rgb(160,160,160);
 $gray: rgb(100,100,100);
 $darkGray: rgb(70,70,70);
@@ -7,9 +8,8 @@ $green: rgb(46,204,64);
 $yellow: rgb(255,255,0);
 $blue: rgb(66,133,244);
 $lightGreen: rgb(45,184,75);
-$themeBlue: rgb(0,11,27);
+$themeBlue: rgb(0,28,73);
 
 // sizing
 $headerHeight: 70px;
 $footerHeight: 100px;
-
