+++ src/shared/types/user.ts
@@ -1,0 +1,4 @@
+export interface User {
+  id: string
+  name: string
+}
