+++ tsconfig.json
@@ -1,9 +1,9 @@
 {
   "compilerOptions": {
     "target": "es5",
+    "module": "esnext",
     "moduleResolution": "node",
     "sourceMap": true,
-    "removeComments": true,
     "noImplicitAny": true,
     "noUnusedLocals": true,
     "outDir": "build",
