+++ tslint.json
@@ -1,6 +1,10 @@
 {
-  "extends": "tslint:recommended",
+  "extends": [
+    "tslint:recommended",
+    "tslint-eslint-rules"
+  ],
   "rules": {
+    "array-bracket-spacing": [true, "never"],
     "forin": false,
     "interface-name": false,
     "max-classes-per-file": false,
@@ -11,6 +15,7 @@
     "no-unused-expression": true,
     "no-shadowed-variable": false,
     "no-string-throw": false,
+    "object-curly-spacing": true,
     "object-literal-sort-keys": false,
     "quotemark": [true, "single"],
     "semicolon": [true, "never"],
